feat : 1차 코드 정리
